Uploading first working version
Currently only supports 3 locations, so it serves mostly as a proof of concept.
This can find a location defined in location_mapping.json, then find the designated arc file, unpack it, find the item lot xml, modify that location with the given ID from input.json, then repack the arc file; Essentially randomizing those locations. The input.json can only currently be done manually, and will require the use of an Archipelago apworld to output it later on.