Update ap_arc.py
Apparently the xml item lot file whenever unpacked and repacked by the arc tool has a tendency to change form a bit. For the time being, I have bandaged the script to...
- Look for an value from the json file named "vanilla_item"
- At the provided line, if itemid is not found at that line, search for the nearest mUnitClass to the provided line that matches vanilla_item
- When found, update the itemid found in the same ClassRef as mUnitClass
- If it does NOT find a value for vanilla_item, instead search for the absolute closest itemid line and change that.

All steps of this also will log when this happens. This should help the majority of any desync problems, but won't be perfect.